requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal== 1.3.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.9.1
boto3==1.34.26
botocore==1.34.26
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
jmespath==1.0.1
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.4
mysql-connector-python==8.3.0
pip==22.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
setuptools==59.6.0
six==1.16.0
urllib3==2.0.7
yarl==1.9.4